% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{access_connect}
\alias{access_connect}
\title{Connect to a Microsoft Access database.}
\usage{
access_connect(path)
}
\arguments{
\item{path}{Path to the Access database.}
}
\value{
A DBI connection.
}
\description{
Connect to a Microsoft Access database.
}
\keyword{internal}
