% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{access_execute_sql}
\alias{access_execute_sql}
\title{Execute SQL queries in a Microsoft Access database.}
\usage{
access_execute_sql(sql, path)
}
\arguments{
\item{sql}{SQL queries as a character vector.}

\item{path}{Path to the Access database.}
}
\description{
Execute SQL queries in a Microsoft Access database.
}
