% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{access_executer_sql}
\alias{access_executer_sql}
\title{Executer des commandes SQL dans une base Access}
\usage{
access_executer_sql(liste_sql, base_access = "Tables_Ref.accdb")
}
\arguments{
\item{liste_sql}{Un vecteur de commandes SQL au format chaîne de caractères.}

\item{base_access}{Chemin de la base Access.}
}
\description{
Exécuter des commandes SQL dans une base Access
}
