% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{access_export}
\alias{access_export}
\title{Export a data frame to a Microsoft Access database (using RODBC).}
\usage{
access_export(data, path, table_name = NULL, override = TRUE)
}
\arguments{
\item{data}{Data frame to export, unquoted.}

\item{path}{Path to the Access database.}

\item{table_name}{Optional name of the table to export as a character. By default, the name of the data frame is used.}

\item{override}{If \code{TRUE} then the new data frame override the Access table if it already exists in the database.}
}
\description{
Export a data frame to a Microsoft Access database (using RODBC).
}
\examples{
impexp::access_export(data_frame, path = "Path/to/the/database.accdb")
impexp::access_export(data_frame, path = "Path/to/the/database.accdb"), table_name = "export")

}
