% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{access_exporter}
\alias{access_exporter}
\title{Exporter une table vers une base Access}
\usage{
access_exporter(table, base_access = "Tables_Ref.accdb",
  table_access = NULL, ecraser = TRUE)
}
\arguments{
\item{table}{Data frame (sans guillemets) à exporter vers Access.}

\item{base_access}{Chemin de la base Access.}

\item{table_access}{Nom de la table Access créée. Par défaut, utilisation du nom du data frame.}

\item{ecraser}{\code{TRUE}, exporter même si une table Access du même nom exite déjà; \code{FALSE}, ne pas écraser si une table du même existe déjà.}
}
\description{
Exporter une table vers une base Access.
}
\examples{
# Export d'un data frame dont le nom dans Access sera "data_frame"
importr::access_exporter(data_frame, base_access = "Chemin/vers/une/base/Access.accdb")

# Export d'un data frame dont le nom dans Access sera "export"
importr::access_exporter(data_frame, base_access = "Chemin/vers/une/base/Access.accdb",
  table_access = "export")

}
