% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{access_import}
\alias{access_import}
\title{Import a Microsoft Access database table.}
\usage{
access_import(table, path)
}
\arguments{
\item{table}{Name of the table to import as a character.}

\item{path}{Path to the Access database.}
}
\value{
A data frame.
}
\description{
Import a Microsoft Access database table.
}
\examples{
impexp::access_import("Table_impexp",
  path = paste0(find.package("impexp"), "/extdata/impexp.accdb"))

}
