% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{access_importer}
\alias{access_importer}
\title{Importer une table Access}
\usage{
access_importer(table, base_access = "Tables_ref.accdb")
}
\arguments{
\item{table}{Nom de la table à importer.}

\item{base_access}{Chemin de la base Access.}
}
\value{
Un data frame correspondant à la table Access.
}
\description{
Importer une table Access.
}
\examples{
impexp::access_importer("Table_impexp",
  base_access = paste0(racine_packages, "impexp/inst/extdata/impexp.accdb"))

}
