% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{access_liste_tables}
\alias{access_liste_tables}
\title{Lister les tables d'une base Access}
\usage{
access_liste_tables(base_access = "Tables_ref.accdb")
}
\arguments{
\item{base_access}{Chemin de la base Access.}
}
\value{
un vecteur de type caractère contenant les noms des tables de la base Access.
}
\description{
Lister les tables d'une base Access.
}
\examples{
impexp::access_liste_tables(paste0(racine_packages, "impexp/inst/extdata/impexp.accdb"))

}
