% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{connexion_access}
\alias{connexion_access}
\title{Se connecter a une base Access}
\usage{
connexion_access(base_access = "Tables_ref.accdb")
}
\arguments{
\item{base_access}{Chemin de la base Access.}
}
\value{
Une connexion à une base Access.
}
\description{
Se connecter à une base Access.
}
\keyword{internal}
