% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odbc.R
\name{connexion_odbc}
\alias{connexion_odbc}
\title{Connexion a une librairie ODBC}
\usage{
connexion_odbc(librairie)
}
\arguments{
\item{librairie}{Librairie ODBC à laquelle se connecter.}
}
\value{
Un objet de type connexion.\cr

Les librairies configurées sont les suivantes :\cr
- \code{bce} : accès à la BCE\cr
- \code{siredo} : accès aux schémas PEDR, IUF, USR\cr
- \code{rnsr} : accès au RNSR\cr
- \code{iuf_2016} : accès à IUF à partir de 2016 (PostgreSQL)\cr

Il est possible de lister les schémas d'une connexion avec la fonction \code{liste_schemas_odbc()}
et les tables d'une connexion avec la fonction \code{liste_tables_odbc()}.
}
\description{
Connexion à une librairie ODBC.
}
\examples{
connexion_bce <- importr::connexion_odbc("bce")

}

