% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{csv_import_path}
\alias{csv_import_path}
\title{Import csv files located in a path.}
\usage{
csv_import_path(path = ".", pattern = "\\\\.csv$", n_csv = Inf,
  parallel = FALSE, zip = FALSE, pattern_zip = "\\\\.zip$",
  progress_bar = FALSE, message = FALSE, ...)
}
\arguments{
\item{path}{Path where the csv files are located (recursive).}

\item{pattern}{A regular expression. Only file names matching the regular expression will be imported.}

\item{n_csv}{Number of csv files to extract. A negative value will starts from the bottom of the files list.}

\item{parallel}{If \code{TRUE} then csv files are imported using all CPU cores (using parallel & pbapply packages).}

\item{zip}{If \code{TRUE} then csv files within zip files are also imported.}

\item{pattern_zip}{A regular expression. Only zip files matching the regular expression will be extracted.}

\item{progress_bar}{If \code{TRUE} then a progress bar is displayed (using pbapply package).}

\item{message}{If \code{TRUE} then a message indicates how many files are imported.}

\item{\dots}{Optional arguments from \code{data.table::fread} function.}
}
\value{
A data frame with a column-list "import" containing all tibbles.
}
\description{
Import csv files located in a path.
}
