% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{excel_export}
\alias{excel_export}
\title{Export a Microsoft Excel file (using openxlsx).}
\usage{
excel_export(
  data,
  path,
  create_dir = FALSE,
  sheet = NULL,
  footer = NULL,
  n_cols_rowname = 0
)
}
\arguments{
\item{data}{A data frame or a named list of data frames that will be sheets in the Excel file.}

\item{path}{Path to the created Excel file.}

\item{create_dir}{if \code{TRUE} then the directory containing the Excel file is created.}

\item{sheet}{Sheet name if data is a data frame, or overrides list names.}

\item{footer}{Footer notes to place beneath tables in sheets.}

\item{n_cols_rowname}{Number of columns containg row names. These columns usualy don't need header.}
}
\description{
Export a Microsoft Excel file (using openxlsx).
}
