% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{excel_exporter}
\alias{excel_exporter}
\title{Exporter un fichier excel}
\usage{
excel_exporter(table, nom_fichier, creer_repertoire = FALSE,
  nom_onglet = NULL, notes = NULL, n_colonnes_lib = 0)
}
\arguments{
\item{table}{data.frame ou liste de data.frame à exporter.}

\item{nom_fichier}{\dots}

\item{creer_repertoire}{\dots}

\item{nom_onglet}{\dots}

\item{notes}{\dots}

\item{n_colonnes_lib}{\dots}
}
\description{
Exporter un fichier excel.
}
