% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{excel_import_path}
\alias{excel_import_path}
\title{Import Microsoft Excel files located in a path.}
\usage{
excel_import_path(path = ".", pattern = "\\\\.xlsx?$",
  pattern_sheet = ".", parallel = FALSE, zip = FALSE,
  progress_bar = TRUE, message = FALSE, ...)
}
\arguments{
\item{path}{Path where the excel files are located (recursive).}

\item{pattern}{A regular expression. Only file names matching the regular expression will be imported.}

\item{pattern_sheet}{A regular expression. Only sheet names in excel files matching the regular expression will be imported.}

\item{parallel}{If \code{TRUE} then excel files are imported using all CPU cores (using parallel package).}

\item{zip}{If \code{TRUE} then excel files within zip files are also imported.}

\item{progress_bar}{If \code{TRUE} then a progress bar is displayed (using pbapply package).}

\item{message}{If \code{TRUE} then a message indicates how many files are imported.}

\item{\dots}{Optional arguments from \code{readxl::read_excel} function.}
}
\value{
A data frame whith a column-list "import" containing all tibbles.
}
\description{
Import Microsoft Excel files located in a path.
}
\examples{
impexp::excel_import_path(paste0(find.package("impexp"), "/extdata"), pattern_sheet = "impexp")
impexp::excel_import_path(paste0(find.package("impexp"), "/extdata"), pattern_sheet = "impexp", skip = 1)
impexp::excel_import_path(paste0(find.package("impexp"), "/extdata"), pattern_sheet = "impexp", zip = TRUE)

}
