% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{excel_importer}
\alias{excel_importer}
\title{Importer un fichier Excel}
\usage{
excel_importer(fichier, nom_onglet = NULL, regex_onglet = NULL,
  num_onglet = 1, ligne_debut = 1, na = NULL, col_types = NULL)
}
\arguments{
\item{fichier}{Chemin vers le fichier excel.}

\item{nom_onglet}{Nom de l'onglet à importer.}

\item{regex_onglet}{Expression régulière à partir de laquelle les onglet dont le nom matche sont importés.}

\item{num_onglet}{Numéro de l'onglet à importer.}

\item{ligne_debut}{Ligne de début à partir duquel importer.}

\item{na}{Caractères à considérer comme vide en plus de \code{c("")}.}

\item{col_types}{Type des champs (utilisé par \code{readxl::read_excel}.}
}
\value{
Un data frame ou une liste de data frames.\cr

Si les paramètres \code{nom_onglet} et \code{regex_onglet} sont laissés à NULL, alors le premier onglet du fichier excel est importé.\cr
Si le paramètre \code{nom_onglet} est rempli, alors cet onglet est importé sans prendre en compte le paramètre \code{regex_onglet}.\cr
Si le paramètre \code{regex_onglet} est rempli et le champ \code{nom_onglet} laissé à vide, alors les onglets dont le nom matche avec l'expression régulière sont importés dans une liste "import" stockée dans un data frame.
}
\description{
Importer un fichier Excel.
}
\examples{
# Import du premier onglet
impexp::excel_importer(paste0(racine_packages, "impexp/inst/extdata/impexp.xlsx"))

# Import de l'onglet "Onglet 2"
impexp::excel_importer(paste0(racine_packages, "impexp/inst/extdata/impexp.xlsx"),
  nom_onglet = "Onglet 2")

# Import des onglets comprenant le mot "Autre" en début de chaine
impexp::excel_importer(paste0(racine_packages, "impexp/inst/extdata/impexp.xlsx"),
  regex_onglet = "^Autre")

}
