% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{excel_importer_masse}
\alias{excel_importer_masse}
\title{Importer les fichiers Excel d'un repertoire (recursif)}
\usage{
excel_importer_masse(regex_fichier, chemin = ".", regex_onglet = ".",
  ligne_debut = 1, na = NULL, col_types = NULL, normaliser = TRUE,
  paralleliser = FALSE, archive_zip = FALSE, message_import = TRUE)
}
\arguments{
\item{regex_fichier}{Expression régulière à partir de laquelle les onglet dont le nom matche sont importés.}

\item{chemin}{Chemin du répertoire à partir duquel seront importés les fichiers excel (récursif).}

\item{regex_onglet}{Expression régulière à partir de laquelle les onglet dont le nom matche sont importés.}

\item{ligne_debut}{Ligne de début à partir duquel importer.}

\item{na}{Caractères à considérer comme vide en plus de \code{c("")}.}

\item{col_types}{Type des champs (utilisé par \code{readxl::read_excel}.}

\item{normaliser}{Normaliser les noms de champ de table.}

\item{paralleliser}{\code{TRUE}, import parallelisé des fichiers excel.}

\item{archive_zip}{\code{TRUE}, les fichiers excel contenus dans des archives zip sont également importés; \code{FALSE} les archives zip sont ignorées.}

\item{message_import}{\code{TRUE}, affichage du message d'import}
}
\value{
Un data frame dont le champ "import" est la liste des data frame importés.
}
\description{
Importer les fichiers Excel d'un répertoire (récursif).
}
\examples{
impexp::importer_masse_xlsx(paste0(find.package("impexp"), "/extdata"), regex_fichier = "xlsx$", regex_onglet = "impexp")

}
