% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{excel_liste_onglets}
\alias{excel_liste_onglets}
\title{Lister les onglets d'un fichier excel}
\usage{
excel_liste_onglets(fichier)
}
\arguments{
\item{fichier}{Chemin vers le fichier excel.}
}
\value{
un vecteur de type caractère contenant les noms des onglets du fichier excel.
}
\description{
Lister les onglets d'un fichier excel.
}
\examples{
impexp::excel_liste_onglets(paste0(racine_packages, "impexp/inst/extdata/impexp.xlsx"))

}
