% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{excel_onglet}
\alias{excel_onglet}
\title{Creer un onglet dans un fichier excel}
\usage{
excel_onglet(classeur, table, nom_onglet, notes = NULL,
  n_colonnes_lib = 0)
}
\arguments{
\item{classeur}{\dots}

\item{table}{\dots}

\item{nom_onglet}{\dots}

\item{notes}{\dots}

\item{n_colonnes_lib}{\dots}
}
\description{
Créer un onglet dans un fichier excel.
}
\keyword{internal}
