% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{excel_sheet}
\alias{excel_sheet}
\title{Create a sheet in a Microsoft Excel file.}
\usage{
excel_sheet(workbook, data, sheet, footer = NULL, n_cols_rowname = 0)
}
\arguments{
\item{workbook}{An \code{openxlsx} worbook object.}

\item{data}{A data frame.}

\item{sheet}{A sheet name.}

\item{footer}{Footer notes to place beneath tables in sheet.}

\item{n_cols_rowname}{Number of columns containg row names. These columns usualy don't need header.}
}
\description{
Create a sheet in a Microsoft Excel file.
}
\keyword{internal}
