% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\name{executer_sql_sqlite}
\alias{executer_sql_sqlite}
\title{Executer des commandes SQL dans une base SQLite}
\usage{
executer_sql_sqlite(liste_sql, base_sqlite)
}
\arguments{
\item{liste_sql}{Un vecteur de commandes SQL au format chaîne de caractères.}

\item{base_sqlite}{Chemin de la base SQLite.}
}
\description{
Exécuter des commandes SQL dans une base SQLite
}
