% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{exporter_fichier_excel}
\alias{exporter_fichier_excel}
\title{Exporter un fichier excel}
\usage{
exporter_fichier_excel(table, nom_fichier, nom_onglet = NULL)
}
\arguments{
\item{table}{data.frame ou liste de data.frame à exporter.}

\item{nom_fichier}{\dots}

\item{nom_onglet}{\dots}
}
\description{
Exporter un fichier excel.
}
