% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\name{exporter_table_sqlite}
\alias{exporter_table_sqlite}
\title{Exporter une table vers une base SQLite}
\usage{
exporter_table_sqlite(table, base_sqlite, nom_table = NULL, ecraser = TRUE)
}
\arguments{
\item{table}{Table à exporter}

\item{base_sqlite}{Chemin vers la bse SQLite.}

\item{nom_table}{Nom de la table à créer dans la base SQLite.}

\item{ecraser}{Ecraser une table du même nom.}
}
\description{
Exporter une table vers une base SQLite
}
