% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{import_test_champ_manquant}
\alias{import_test_champ_manquant}
\title{Importer un fichier Excel sans connaitre la ligne de début, mais à partir de la première ligne non-vide d'un champ (fonction générique).}
\usage{
import_test_champ_manquant(table, test_champ_manquant, fichier, num_onglet,
  col_types)
}
\arguments{
\item{table}{Table pour laquelle doit être déterminée la première ligne non-vide.}

\item{test_champ_manquant}{Un nom de champ du fichier. Importer un fichier Excel sans connaitre la ligne de début, mais à partir de la première ligne non-vide du champ.}

\item{fichier}{Chemin vers le fichier excel.}

\item{num_onglet}{Numéro de l'onglet à importer.}

\item{col_types}{Type des champs (utilisé par \code{readxl::read_excel}.}
}
\value{
Un data frame importé à partir de la première ligne non-vide.
}
\description{
Importer un fichier Excel sans connaitre la ligne de début, mais à partir de la première ligne non-vide d'un champ (fonction générique).
}
\keyword{internal}
