% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odbc.R
\name{importer_base_odbc}
\alias{importer_base_odbc}
\title{Importer une base ODBC}
\usage{
importer_base_odbc(librairie, schema = NULL, message_table = TRUE,
  fichier_sauvegarde = NULL, copie_rdata_date = FALSE)
}
\arguments{
\item{librairie}{Librairie ODBC à laquelle se connecter.}

\item{schema}{Nom du schéma ODBC avec lequel filtrer la liste des tables importées.}

\item{message_table}{Affichage d'un message dans la console lors de l'import de la table.}

\item{fichier_sauvegarde}{Chemin du fichier RData de sauvegarde.}

\item{copie_rdata_date}{\code{TRUE}, copie datée de la base pour archive; \code{FALSE}, aucune copie n'est réalisée.}
}
\value{
Une liste nommée de data frame correspondant aux tables de la base ODBC.
}
\description{
Importer une base ODBC.
}
\examples{
# Import de la base IUF (2015 et avant)
iuf <- importr::importer_base_odbc(librairie = "siredo", schema = "IUF")
iuf

}

