% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{importer_fichier_csv}
\alias{importer_fichier_csv}
\title{Importer un fichier CSV}
\usage{
importer_fichier_csv(fichier, ligne_debut = 1, encoding = "Latin-1",
  col_types = NULL)
}
\arguments{
\item{fichier}{Chemin vers le fichier CSV}

\item{ligne_debut}{Ligne de début à partir duquel importer.}

\item{encoding}{Encodage du fichier CSV.}

\item{col_types}{Type des champs (utilisé par \code{data.table::fread}.}
}
\value{
Un data frame.\cr
}
\description{
Importer un fichier CSV.
}
