% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{importer_masse_excel}
\alias{importer_masse_excel}
\title{Importer les fichiers Excel d'un repertoire (recursif)}
\usage{
importer_masse_excel(regex_fichier, chemin = ".", regex_onglet = ".",
  ligne_debut = 1, col_types = NULL, paralleliser = FALSE,
  archive_zip = FALSE, test_champ_manquant = NULL,
  archive_zip_repertoire_sortie = "import_masse_excel",
  message_import = "Import des fichiers excel:")
}
\arguments{
\item{regex_fichier}{Expression régulière à partir de laquelle les onglet dont le nom matche sont importés.}

\item{chemin}{Chemin du répertoire à partir duquel seront importés les fichiers excel (récursif).}

\item{regex_onglet}{Expression régulière à partir de laquelle les onglet dont le nom matche sont importés.}

\item{ligne_debut}{Ligne de début à partir duquel importer.}

\item{col_types}{Type des champs (utilisé par \code{readxl::read_excel}.}

\item{paralleliser}{\code{TRUE}, import parallelisé des fichiers excel.}

\item{archive_zip}{\code{TRUE}, les fichiers excel contenus dans des archives zip sont également importés; \code{FALSE} les archives zip sont ignorées.}

\item{test_champ_manquant}{Un nom de champ du fichier. Importer un fichier Excel sans connaitre la ligne de début, mais à partir de la première ligne non-vide du champ.}

\item{archive_zip_repertoire_sortie}{Nom du répertoire d'extraction des archives zip.}

\item{message_import}{Le message à afficher pendant l'importation (par défaut : "Import des fichiers excels:")}
}
\value{
Un data frame dont le champ "import" est la liste des data frame importés.
}
\description{
Importer les fichiers Excel d'un répertoire (récursif).
}
\examples{
importr::importer_masse_xlsx(paste0(racine_packages, "importr/inst/extdata"), regex_fichier = "xlsx$", regex_onglet = "importr")

}
