% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{importer_table_access}
\alias{importer_table_access}
\title{Importer une table Access}
\usage{
importer_table_access(table, base_access = "Tables_ref.accdb")
}
\arguments{
\item{table}{Nom de la table à importer.}

\item{base_access}{Chemin de la base Access.}
}
\value{
Un data frame correspondant à la table Access.
}
\description{
Importer une table Access.
}
\examples{
importr::importer_table_access("Table_importr",
  base_access = paste0(racine_packages, "importr/inst/extdata/importr.accdb"))

}
