% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libreoffice.R
\name{importer_table_odb}
\alias{importer_table_odb}
\title{Importer une table LibreOffice}
\usage{
importer_table_odb(table, base_odb = "Tables_ref.odb")
}
\arguments{
\item{table}{Nom de la table à importer.}

\item{base_odb}{Chemin de la base LibreOffice.}
}
\value{
Un data frame correspondant à la table LibreOffice.\cr

Ne fonctionne pas sous Windows.
}
\description{
Importer une table LibreOffice.
}
\examples{
importr::importer_table_odb("importr",
  base_odb = paste0(racine_packages, "importr/inst/extdata/importr.odb"))

}

