% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odbc.R
\name{importer_table_odbc}
\alias{importer_table_odbc}
\title{Importer une table d'une base ODBC}
\usage{
importer_table_odbc(table, connexion_odbc, schema = NULL,
  message_table = FALSE)
}
\arguments{
\item{table}{Nom de table ODBC à importer.}

\item{connexion_odbc}{Connexion ODBC établie avec la fonction \code{connexion_odbc()}.}

\item{schema}{Nom du schéma ODBC correspondant.}

\item{message_table}{Affichage d'un message dans la console lors de l'import de la table.}
}
\value{
Un data frame correspondant à la table ODBC.\cr

L'import est plus rapide si le paramètre \code{schema} est rempli.
}
\description{
Importer une table d'une base ODBC.
}
\examples{
# Import de la table "N_NATURE_UAI" de la BCE
connexion_bce <- importr::connexion_odbc("bce")
n_nature_uai <- importr::importer_table_odbc("N_NATURE_UAI", connexion_bce, schema = "BCE")

}

