% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas.R
\name{importer_table_sas}
\alias{importer_table_sas}
\title{Importer une table SAS}
\usage{
importer_table_sas(fichier)
}
\arguments{
\item{fichier}{Chemin de table sas.}
}
\value{
Un data frame correspondant à la table SAS.
}
\description{
Importer une table SAS
}
\examples{
importr::importer_table_sas(paste0(racine_packages, "importr/inst/extdata/importr.sas7bdat"))

}

