% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\name{importer_table_sqlite}
\alias{importer_table_sqlite}
\title{Importer une table d'une base SQLite}
\usage{
importer_table_sqlite(table, base_sqlite)
}
\arguments{
\item{table}{Nom de la table à importer.}

\item{base_sqlite}{Chemin vers la bse SQLite.}
}
\value{
Un tibble.
}
\description{
Importer une table d'une base SQLite
}
