% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{liste_onglets_excel}
\alias{liste_onglets_excel}
\title{Lister les onglets d'un fichier excel}
\usage{
liste_onglets_excel(fichier)
}
\arguments{
\item{fichier}{Chemin vers le fichier excel.}
}
\value{
un vecteur de type caractère contenant les noms des onglets du fichier excel.
}
\description{
Lister les onglets d'un fichier excel.
}
\examples{
importr::liste_onglets_excel(paste0(racine_packages, "importr/inst/extdata/importr.xlsx"))

}

