% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odbc.R
\name{liste_schemas_odbc}
\alias{liste_schemas_odbc}
\title{Lister les schemas d'une base ODBC}
\usage{
liste_schemas_odbc(connexion_odbc)
}
\arguments{
\item{connexion_odbc}{Connexion ODBC établie avec la fonction \code{connexion_odbc()}.}
}
\value{
Un vecteur des schémas ODBC de la connexion.
}
\description{
Lister les schémas d'une base ODBC.
}
\examples{
# La liste des schémas de la BCE
importr::connexion_odbc("bce") \%>\%
  importr::liste_schemas_odbc()

}

