% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{liste_tables_access}
\alias{liste_tables_access}
\title{Lister les tables d'une base Access}
\usage{
liste_tables_access(base_access = "Tables_ref.accdb")
}
\arguments{
\item{base_access}{Chemin de la base Access.}
}
\value{
un vecteur de type caractère contenant les noms des tables de la base Access.
}
\description{
Lister les tables d'une base Access.
}
\examples{
importr::liste_tables_access(paste0(racine_packages, "importr/inst/extdata/importr.accdb"))

}
