% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importr.R
\name{normaliser_nom_champs}
\alias{normaliser_nom_champs}
\title{Normaliser les noms des champs d'une table importee}
\usage{
normaliser_nom_champs(table)
}
\arguments{
\item{table}{Un data frame.}
}
\value{
Un data frame avec les noms de champ sont normalisés.\cr

La fonction \code{caractr::normaliser_char()} est appliquée sur chacun des noms de champ.
}
\description{
Normaliser les noms des champs d'une table importée.
}
\examples{
table <- dplyr::data_frame(
  "Type d'unité Sirus : entreprise profilée ou unité légale" = NA_character_,
  "Nic du siège"= NA_character_
)
importr::normaliser_nom_champs(table)

}
