% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdata.R
\name{r_import}
\alias{r_import}
\title{Import an objet from a R data file.}
\usage{
r_import(r_data, object = NULL)
}
\arguments{
\item{r_data}{R data file path.}

\item{object}{Object name to load.}
}
\value{
If parameter \code{object} stays \code{NULL} then all objects of the R data file loaded in a named list.\cr
Otherwise, only the object specified is loaded.
}
\description{
Import an objet from a R data file.
}
\examples{
impexp::r_import(system.file("extdata/impexp.RData", package = "impexp"))
impexp::r_import(system.file("extdata/impexp.RData", package = "impexp"), "data")

}
