% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas.R
\name{sas_importer}
\alias{sas_importer}
\title{Importer une table SAS}
\usage{
sas_importer(fichier)
}
\arguments{
\item{fichier}{Chemin de table sas.}
}
\value{
Un data frame correspondant à la table SAS.
}
\description{
Importer une table SAS
}
\examples{
impexp::sas_importer(paste0(racine_packages, "impexp/inst/extdata/impexp.sas7bdat"))

}
