% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\name{sqlite_creer}
\alias{sqlite_creer}
\title{Creer une base SQLite vide}
\usage{
sqlite_creer(base_sqlite)
}
\arguments{
\item{base_sqlite}{Chemin de la base SQLite.}
}
\description{
Créer une base SQLite vide
}
