% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\name{sqlite_executer_sql}
\alias{sqlite_executer_sql}
\title{Executer des commandes SQL dans une base SQLite}
\usage{
sqlite_executer_sql(liste_sql, base_sqlite)
}
\arguments{
\item{liste_sql}{Un vecteur de commandes SQL au format chaîne de caractères.}

\item{base_sqlite}{Chemin de la base SQLite.}
}
\description{
Exécuter des commandes SQL dans une base SQLite
}
