% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\name{sqlite_exporter}
\alias{sqlite_exporter}
\title{Exporter une table vers une base SQLite}
\usage{
sqlite_exporter(table, base_sqlite, nom_table = NULL, ecraser = TRUE)
}
\arguments{
\item{table}{Table à exporter}

\item{base_sqlite}{Chemin de la base SQLite.}

\item{nom_table}{Nom de la table à créer dans la base SQLite.}

\item{ecraser}{Ecraser une table du même nom.}
}
\description{
Exporter une table vers une base SQLite
}
