% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\name{sqlite_import}
\alias{sqlite_import}
\title{Import a SQLite database table.}
\usage{
sqlite_import(table, path)
}
\arguments{
\item{table}{Name of the table to import as a character.}

\item{path}{SQLite database path.}
}
\value{
A tibble.
}
\description{
Import a SQLite database table.
}
