% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\name{sqlite_list_tables}
\alias{sqlite_list_tables}
\title{Return table names from a SQLite database.}
\usage{
sqlite_list_tables(path)
}
\arguments{
\item{path}{SQLite database path.}
}
\description{
Return table names from a SQLite database.
}
