% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\name{sqlite_liste_tables}
\alias{sqlite_liste_tables}
\title{Liste les tables d'une base SQLite}
\usage{
sqlite_liste_tables(base_sqlite)
}
\arguments{
\item{base_sqlite}{Chemin de la base SQLite.}
}
\description{
Liste les tables d'une base SQLite
}
