% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh.R
\name{ssh_import_file}
\alias{ssh_import_file}
\title{ssh_import_file}
\usage{
ssh_import_file(path, ssh_key_file, to = ".")
}
\arguments{
\item{path}{\dots}

\item{ssh_key_file}{\dots}

\item{to}{\dots}
}
\description{
ssh_import_file
}
