% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\name{ajouter_lignes_sqlite}
\alias{ajouter_lignes_sqlite}
\title{Ajouter des lignes a une table SQLite}
\usage{
ajouter_lignes_sqlite(table_ajout, table, base_sqlite)
}
\arguments{
\item{table_ajout}{Table à concaténer.}

\item{table}{Table initiale.}

\item{base_sqlite}{Chemin de la base SQLite.}
}
\description{
Ajouter des lignes à une table SQLite
}
