% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importr.R
\name{caracteres_vides_na}
\alias{caracteres_vides_na}
\title{Remplacer les NA par des caracteres vides}
\usage{
caracteres_vides_na(table)
}
\arguments{
\item{table}{Un data frame.}
}
\value{
Un data frame dont les champ à \code{""} sont transformés en \code{NA}.
}
\description{
Remplacer les NA par des caractères vides.
}
\examples{
table <- dplyr::data_frame(
  champ_1 = "",
  champ_2 = ""
)
importr::caracteres_vides_na(table)

}

