% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{creer_onglet_excel}
\alias{creer_onglet_excel}
\title{Creer un onglet dans un fichier excel}
\usage{
creer_onglet_excel(classeur, table, nom_onglet)
}
\arguments{
\item{classeur}{\dots}

\item{table}{\dots}

\item{nom_onglet}{\dots}
}
\description{
Créer un onglet dans un fichier excel.
}
\keyword{internal}
