% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{excel_importer_}
\alias{excel_importer_}
\title{Importer un fichier Excel (fonction générique).}
\usage{
excel_importer_(fichier, num_onglet, ligne_debut = 1, na = NULL,
  col_types = NULL)
}
\arguments{
\item{fichier}{Chemin vers le fichier excel.}

\item{num_onglet}{Numéro de l'onglet à importer.}

\item{ligne_debut}{Ligne de début à partir duquel importer.}

\item{na}{Caractères à considérer comme vide en plus de \code{c("")}.}

\item{col_types}{Type des champs (utilisé par \code{readxl::read_excel}.}
}
\value{
Un data frame correspondant à la feuille excel.
}
\description{
Importer un fichier Excel (fonction générique).
}
\keyword{internal}
