% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{importer_fichier_csv}
\alias{importer_fichier_csv}
\title{Importer un fichier CSV}
\usage{
importer_fichier_csv(fichier, ligne_debut = 1, encoding = "Latin-1",
  na = NULL, col_types = NULL, warning_type = TRUE)
}
\arguments{
\item{fichier}{Chemin vers le fichier CSV}

\item{ligne_debut}{Ligne de début à partir duquel importer.}

\item{encoding}{Encodage du fichier CSV.}

\item{na}{Caractères à considérer comme vide en plus de \code{c("NA", "", " ")}.}

\item{col_types}{Type des champs (utilisé par \code{data.table::fread}.}

\item{warning_type}{Affichage des warnings liés aux types des champs importés.}
}
\value{
Un data frame.\cr
}
\description{
Importer un fichier CSV.
}
