% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdata.R
\name{rdata_import}
\alias{rdata_import}
\title{Import an objet from a RData file.}
\usage{
rdata_import(rdata, object = NULL)
}
\arguments{
\item{rdata}{RData file path.}

\item{object}{Object name to load.}
}
\value{
If parameter \code{object} stays \code{NULL} then all objects of the RData file loaded in a named list.\cr
Otherwise, only the object specified is loaded.
}
\description{
Import an objet from a RData file.
}
\examples{
impexp::rdata_import(paste0(find.package("impexp"), "/extdata/impexp.RData"))
impexp::rdata_import(paste0(find.package("impexp"), "/extdata/impexp.RData"), "data")

}
