% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\name{sqlite_ajouter_lignes}
\alias{sqlite_ajouter_lignes}
\title{Ajouter des lignes a une table SQLite}
\usage{
sqlite_ajouter_lignes(table_ajout, table, base_sqlite,
  attente_verrou = TRUE)
}
\arguments{
\item{table_ajout}{Table à concaténer.}

\item{table}{Table initiale.}

\item{base_sqlite}{Chemin de la base SQLite.}

\item{attente_verrou}{Tentatives répétées jusqu'à ce que le verrou soit libéré.}
}
\description{
Ajouter des lignes à une table SQLite
}
\details{
L'ordre des champs de la table à ajouter doit suivre celui de la table initiale
}
