% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\name{sqlite_append_rows}
\alias{sqlite_append_rows}
\title{Append rows to a table in a SQLite database.}
\usage{
sqlite_append_rows(path, data, table_name = NULL)
}
\arguments{
\item{path}{SQLite database path.}

\item{data}{Data frame to append, unquoted.}

\item{table_name}{Table name in SQLite database to append data.}
}
\description{
Append rows to a table in a SQLite database.
}
