% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\name{sqlite_export}
\alias{sqlite_export}
\title{Export a table to a SQLite database.}
\usage{
sqlite_export(path, table, table_name = NULL, wait_unlock = TRUE, ...)
}
\arguments{
\item{path}{SQLite database path.}

\item{table}{Data frame to export, unquoted.}

\item{table_name}{Optional name of the table to export as a character. By default, the name of sqlite database is used.}

\item{wait_unlock}{Wait until SQLite database is unlocked.}

\item{\dots}{Additional arguments to \code{DBI::dbWriteTable}.}
}
\description{
Export a table to a SQLite database.
}
