% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\name{sqlite_export}
\alias{sqlite_export}
\title{Export a table to a SQLite database.}
\usage{
sqlite_export(table, path, table_name = NULL, override = FALSE,
  message = TRUE)
}
\arguments{
\item{path}{SQLite database path.}

\item{table_name}{Optional name of the table to export as a character. By default, the name of the data frame is used.}

\item{override}{If \code{TRUE} then the new data frame override the SQLite table if it already exists in the database.}

\item{data}{Data frame to export, unquoted.}
}
\description{
Export a table to a SQLite database.
}
