% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.R
\name{sqlite_import}
\alias{sqlite_import}
\title{Import a SQLite database table.}
\usage{
sqlite_import(path, table = NULL, wait_unlock = TRUE, ...)
}
\arguments{
\item{path}{SQLite database path.}

\item{table}{Name of the table to import as a character.}

\item{wait_unlock}{Wait until SQLite database is unlocked.}

\item{\dots}{Additional arguments passed to \code{DBI::dbReadTable}.}
}
\value{
A tibble.
}
\description{
Import a SQLite database table.
}
